app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "notebook>=7.5.5",
    "numpy>=2.4.4",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
